Separation concerns. Array and value helpers added.
