+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tiberium-flat",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "description": "",
   "main": "index.js",
   "scripts": {
