+++ values.js
@@ -1,0 +1,7 @@
+const valuesTiberiumHelper = {
+  returnValue: function(val) {
+    return val;
+  }
+};
+
+module.exports = valuesTiberiumHelper;
